+++ src/me/hkgumbs/main/java/tictactoe/Minimax.java
@@ -1,4 +1,4 @@
-package main.java;
+package me.hkgumbs.main.java.tictactoe;
 
 import java.util.ArrayList;
 import java.util.Collections;
